+++ src/styles/About.css
@@ -63,6 +63,13 @@
         font-size: 1.3em;
     }
 }
+
+@media (min-height: 1300px) { 
+    .About img {
+        display: none;
+    }
+}
+
 /*.Life {
     margin-left: auto;
     margin-right: auto;
