+++ src/styles/Work.css
@@ -1,3 +1,161 @@
+.Work .left .chatbox {
+    background-color: #2f2f2f;
+    color:white;
+    padding:7px;
+    border-radius:5px;
+    display: inline-block;
+}
+
+.Work .right .chatbox {
+    background-color: aquamarine;
+    color:#2f2f2f;
+    padding:7px;
+    border-radius:5px;
+    display: inline-block;
+}
+
+.Work p {
+    margin-top: 10px;
+    margin-bottom:10px;
+}
+
+.Work {
+    margin-top: 20px;
+    width: 100%;
+    max-width: 800px;
+    display: block;
+    margin-left:auto;
+    margin-right: auto
+}
+
+.Work p.right {
+    text-align: right;
+}
+
+.left + .right, .right + .left {
+    margin-top: 30px;
+}
+
+/*Errors*/
+
+.errors {
+    font-size: .8em;
+    margin-top: 6px;
+    font-weight: bold;
+    height: 20px;
+    text-align: center;
+    font-weight: bold;
+}
+
+/*Input number*/
+
+.Work input[type=number]:focus{
+    outline-width: 0;
+    outline:none;
+}
+
+.Work input[type=number]::-webkit-inner-spin-button, 
+.Work input[type=number]::-webkit-outer-spin-button { 
+    -webkit-appearance: none;
+    -moz-appearance: none;
+    appearance: none;
+    margin: 0; 
+}
+
+.Work input[type="number"] {
+    -moz-appearance: textfield;
+    width: 50px;
+    text-align: center;
+    vertical-align: center;
+    border: 2px transparent solid;
+    border-radius: 7px;
+    font-size: inherit;
+    margin-right: 10px;
+}
+
+/*Checkboxes*/
+
+@keyframes tick {
+    0% {transform: scale(0)}
+    90% {transform: scale(1.3)}
+    100% {transform: scale(1)}
+}
+
+.Work label {
+    display: block;
+    text-align:left
+}
+
+.Work .label-item:before {
+    font-family: 'FontAwesome'; 
+    margin-right: 5px;   
+    display: inline-block;
+    width: 1em;
+}
+
+.Work input[type="checkbox"]:not(checked) + label .label-item:before {
+    font-family: 'FontAwesome';
+    content:"\f096";
+}
+
+.Work input[type="checkbox"]:checked + label .label-item:before {
+    content: "\f14a";
+    animation-name: tick;
+    animation-duration: 90ms;
+    animation-timing-function: ease-in;
+}
+
+.Work input[type="checkbox"] {
+    display: none;
+}
+
+.Work button {
+    background-color: #2f2f2f;
+    color:white;
+    padding:7px;
+    border-radius:5px;
+    display: inline-block;
+    font-size: inherit;
+    user-select: none;
+    cursor: pointer;
+    margin-left: auto;
+    margin-right: auto;
+    display: block;
+    margin-top: 30px;
+    border: 1px transparent solid;
+}
+
+.Work button:focus {
+    outline-width: 0;
+}
+
+.Work button:hover {
+    background-color: aquamarine;
+    color:#2f2f2f;
+    transition: background-color 0.3s ease;
+}
+
+@media (max-width: 576px) { 
+    .Work, .Work .Work input[type=number] {
+        font-size: 16px;
+    }
+    .Work .right {
+        font-size: 20px
+    }
+    .Work button:hover {
+        color: white;
+        background-color: #2f2f2f;
+    }
+}
+
+@media only screen 
+and (max-device-width : 1024px)  { 
+    .Work button:hover {
+        color: white;
+        background-color: #2f2f2f;
+    }
+}
+
 /*.Work {
     display: block;
     width: 670px;
